third_party/blink/web_tests/external/wpt/mediacapture-region/CropTarget-fromElement.https.html

<!doctype html>
<html>

<head>
  <title>Test CropTarget.fromElement()</title>
  <meta name='assert' content='Test CropTarget.fromElement().' />
</head>

<body>
  <h1 class="instructions">Description</h1>
  <p class="instructions">
    This test checks for the behavior of <code>CropTarget.fromElement()</code>.
  </p>

  <div id='test-div'></div>
  <iframe id='test-iframe' src="about:blank"></iframe>
  <a id='test-a'></a>
  <div id='log'></div>

  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>

  <script>
    "use strict";

    // Regex that matches a Universally Unique IDentifer composed of hex
    // characters separated by dashes in the form 8-4-4-4-12 for a total
    // of 36 characters.
    const GUID_REGEX = /^[0-9A-Fa-f]{8}(?:-[0-9A-Fa-f]{4}){3}-[0-9A-Fa-f]{12}$/;

    promise_test(async () => {
      assert_true(!!CropTarget.fromElement);
      const iframe_id = await CropTarget.fromElement(
        document.getElementById('test-iframe'));
      assert_true(GUID_REGEX.test(iframe_id));
    }, "produces valid IDs for iframe");

    promise_test(async () => {
      assert_true(!!CropTarget.fromElement);
      const div_id = await CropTarget.fromElement(
        document.getElementById('test-div'));
      assert_true(GUID_REGEX.test(div_id));
    }, "produces valid id for div");

    promise_test(async () => {
      assert_true(!!CropTarget.fromElement);
      const iframe_id = await CropTarget.fromElement(
        document.getElementById('test-iframe'));
      const second_iframe_id = await CropTarget.fromElement(
        document.getElementById('test-iframe'));
      assert_equals(iframe_id, second_iframe_id);
    }, "repeated calls return the same value");

    promise_test(t => {
      assert_true(!!CropTarget.fromElement);
      return promise_rejects_js(t, TypeError,
        CropTarget.fromElement(undefined));
    }, "rejects undefined with a TypeError");

    promise_test(t => {
      assert_true(!!CropTarget.fromElement);
      return promise_rejects_js(t, TypeError, CropTarget.fromElement(123));
    }, "rejects a non-element with a TypeError");

    promise_test(function (t) {
      assert_true(!!CropTarget.fromElement);
      return promise_rejects_dom(t, "NotSupportedError",
        CropTarget.fromElement(document.getElementById("test-a")));
    }, "rejects unimplemented Element subtypes with a NotSupportedError");

    promise_test(async () => {
      assert_true(!!CropTarget.fromElement);
      const div_id = await CropTarget.fromElement(
        document.getElementById('test-div'));
      const iframe_id = await CropTarget.fromElement(
        document.getElementById('test-iframe'));
      assert_not_equals(div_id, iframe_id);
    }, "two elements have different IDs");

    promise_test(async () => {
      assert_true(!!CropTarget.fromElement);
      const div = document.getElementById('test-div');
      const div_id = await CropTarget.fromElement(div);
      const clone = div.cloneNode(true);
      document.querySelector('body').appendChild(clone);
      const clone_id = await CropTarget.fromElement(clone);
      assert_not_equals(div_id, clone_id);
    }, "cloned elements have different IDs");

  </script>
</body>

</html>
